refactor: makes pipeline happy
import methods from other file may cause
some problem when use fate_test.
simply copy common methods to all examples using them.
