--- examples/pipeline/homo_nn/__init__.py
@@ -1,6 +1,0 @@
-import os
-import sys
-
-additional_path = os.path.realpath('../')
-if additional_path not in sys.path:
-    sys.path.append(additional_path)
